Overflows introduced

+++ chart_painter.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "misc.h"
 
-std::unique_ptr<QWidget> create_chart (std::vector<std::pair<double, int>> values_and_case_counts, std::string name);
+std::unique_ptr<QWidget> create_chart (std::vector<std::pair<double, uint64_t>> values_and_case_counts, std::string name);
